Update: MetricReporter: run id values must be int

+++ src/main/java/org/opennars/lab/metric/MetricReporter.java
@@ -11,7 +11,7 @@ import java.util.*;
  */
 public class MetricReporter {
     public String narsVersion = Nar.VERSION;
-    public long runId =  Calendar.getInstance().getTimeInMillis();
+    public int runId = (int)Calendar.getInstance().getTimeInMillis();
 
     public List<MetricObserver> observers = new ArrayList<>();
